+++ calculator/MultiplicationNode.java
@@ -6,7 +6,7 @@ public class MultiplicationNode extends BinaryNode {
     } // end constructor
 
     @Override
-    public int evaluate(int n1, int n2) {
+    public double evaluate(double n1, double n2) {
         return n1 * n2;
     }
 } // end MultiplicationNode
