+++ calculator/Node.java
@@ -7,5 +7,5 @@ package calculator;
 public interface Node {
     // Used to traverse the tree.
     // Evaluates each node in the tree.
-    int evaluate();
+    double evaluate();
 } // end Node
